Disappearing Step removed.
Issue was occurring with saved MoveSequences, wherein they would be
usable once and only once. The issue was related to the fact that a
reference to a single unique MoveSequence was being passed around, and
after being used once, the index was not reset and therefore the move
appeared to have no “Next”. A Reset() method has been added and is
called before attempting to use any move inside of CharacterData.